main_hbt/requirements.txt
absl-py==2.2.2
amqp==5.3.1
asgiref==3.8.1
attrs==25.3.0
billiard==4.2.1
celery==5.4.0
#cffi==1.17.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
dj-database-url==2.3.0
Django==5.1.3
django-cors-headers==4.7.0
django-enumfields==2.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
flatbuffers==25.2.10
fonttools==4.57.0
#jax==0.5.3
#jaxlib==0.5.3
kiwisolver==1.4.8
kombu==5.4.2
#matplotlib==3.10.1
#mediapipe==0.10.21
#ml_dtypes==0.5.1
# mysqlclient==2.2.6  if you use local mysql workbench must installed
numpy==1.26.4
#opencv-contrib-python==4.11.0.86
#opencv-python==4.11.0.86
#opt_einsum==3.4.0
packaging==24.2
pandas==2.2.3
# pillow==11.1.0
prompt_toolkit==3.0.50
protobuf==4.25.6
psycopg==3.2.9
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.1
redis==5.2.1
#scipy==1.15.2
#sentencepiece==0.2.0
six==1.17.0
#sounddevice==0.5.1
sqlparse==0.5.2
typing_extensions==4.13.2
tzdata==2024.2
vine==5.1.0
wcwidth==0.2.13
